requirements.txt
fastapi==0.110.2
pydantic-settings==2.2.1
requests==2.32.4
telnetlib3==2.0.8
uvicorn[standard]==0.29.0
python-dotenv==1.1.1
openai==2.15.0